Up to init_mines_around (function is midway done)

+++ src/headers/main.h
@@ -9,7 +9,7 @@
 int const ROWS = 9;
 int const COLUMNS = 9;
 int const MINES = 10;
-float const CELL_SIZE = 60;
+int const CELL_SIZE = 60;
 
 // Dimensions de la fenêtre
 int const WIN_WIDTH = ROWS * CELL_SIZE;
